Calculate FID scores during training
This commit adds the calculation of FID scores during the training process. It leverages the `calculate_fid_scores` script and integrates it into the `__sample_during_training` method. The FID scores are computed using the validation images specified in the `concepts.json` file and the generated samples saved in the hidden "epochs" directory.