Image preview from subdirectories, filename/caption preview
Image preview for concepts now looks in subdirectories if option is enabled, in both the concept tab and concept window. Tested with files in the top level and nested subdirectories and works as expected, no noticeable lag when previewing a concept folder with several thousand images.

Adds fields in concept window to display the name of the previewed file and the associated caption, depending on prompt source. Currently only displays the "unmodified" caption before shuffling/dropout, and only the first line if a file has multiple captions. Will try to change this to preview the "modified" caption in the future.