Concept statistics
Add new tab in concept window which will display information about the file contents of the concept. Will show the total size of all files and the number of images, masks, and captions. Will also check how many images are paired with a caption/mask of the same name, and conversely if there are any captions/masks which lack a corresponding image.

To do:
- currently only runs when "refresh" button is clicked, as large datasets can take a long time to process and freeze the program. Will try to have it start running on opening the window with a timeout of ~1/2 second, so smaller concepts can be processed automatically but larger ones won't hold the program
- save/load stats to the concept config file so they can be retrieved more quickly, assuming no files have been added/deleted since it was last run
- better formatting on the tab to make labels/data more clear
- more statistics on image resolution, tag frequency, or other useful information. Some of this may take a lot longer to process so parts may be enabled/disabled depending on user input