multithreading
rearranged functions and added multithreading with concurrent.futures, with each video to extract from or link to download in a new thread. Seems to speed up the process significantly when dealing with multiple videos, may still need to test it with a larger number of videos or with a longer video to ensure no excessive memory usage or slowdown.

Added some more info to tooltips on GUI, and a link to the yt-dlp github for reference on arguments and compatibility. Commented out the feature detection options until I do more testing with cv2's classifiers and figure out how it should work from a user perspective