concept stats scan abort/timeout performance
Add a faster check for stopping the concept scan when it hits the timeout limit or is manually set to stop, using threading event passed to function

Add "subcaption" count to advanced stats, which is the number of captions in each line across all files. Also displayed in parentheses in the ui. May change the variable name later

Some WIP experiments with a parallel version of the subfolder scan using concurrent.futures, scanning each folder in a separate thread then combining them together at the end. Faster for files on my HDD array but slower on SSD so need to optimize it before making it default